// backend/src/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  FRIEND_REQUEST = 'FRIEND_REQUEST',
  GAME_INVITE = 'GAME_INVITE',
  GAME_INVITE_ACCEPTED = 'GAME_INVITE_ACCEPTED',
  GAME_ABORTED = 'GAME_ABORTED',
  TOURNAMENT_ALERT = 'TOURNAMENT_ALERT',
  TOURNAMENT_REMINDER = 'TOURNAMENT_REMINDER',
  CLUB_ROLE_UPDATE = 'CLUB_ROLE_UPDATE',
  CLUB_MEMBER_JOINED = 'CLUB_MEMBER_JOINED',
  CLUB_MEMBER_LEFT = 'CLUB_MEMBER_LEFT',
  SUPER_ADMIN_TRANSFER = 'SUPER_ADMIN_TRANSFER',
  GAME_RESULT = 'GAME_RESULT',
  MATCH_READY = 'MATCH_READY',
  SYSTEM_ALERT = 'SYSTEM_ALERT',
} 